requirements.txt
# This file contains all required packages for both the Streamlit dashboard and Kafka producer
# Core packages for producer and dashboard
streamlit>=1.28.0              # Web dashboard framework
pandas>=2.0.0                  # Data manipulation
kafka-python>=2.0.2            # Kafka producer/consumer
pymongo>=4.5.0                 # MongoDB integration
requests>=2.31.0               # API calls to WeatherAPI
python-dateutil>=2.8.2         # Date/time parsing

# Optional: Auto-refresh for Streamlit
streamlit-autorefresh>=0.0.3   # Enables auto-refresh behavior

# Optional: Visualization enhancements
plotly>=5.15.0                 # Interactive charts (if used)
matplotlib>=3.7.0              # Static plots (if used)

# Optional: Development tools
pytest>=7.4.0                  # Unit testing
black>=23.0.0                  # Code formatting
flake8>=6.0.0                  # Linting

# Optional: Machine learning (if extended later)
# scikit-learn>=1.3.0
# scipy>=1.10.0

# Optional: HDFS support (if you switch from MongoDB)
# pyarrow>=12.0.0

# Notes:
# - MongoDB is used for storage in this project.
# - WeatherAPI requires a valid API key.
# - Kafka must be running locally on port 9092.
